+++ CHANGES.md
@@ -8,6 +8,7 @@
 - test(pkg:array): added unit tests for `Deduplicate` function
 - fix(pkg:async): added `destroy` method to `AbortTimeoutController` to release resources in time
 - fix(pkg:async): modified `abort` method of `AbortTimeoutController` to prevent aborting multiple times
+- feat(pkg:async): added API `Async.withAbortSignal`
 
 ## v1.5.0